ccgx: Split into ccgx-dmc
This was really two plugins in a trenchcoat.
