+++ plugins/ccgx-dmc/fu-ccgx-dmc-device.h
