+++ plugins/ccgx-dmc/fu-ccgx-dmc-devx-device.h
