+++ plugins/ccgx-dmc/fu-ccgx-dmc-firmware.c
