+++ plugins/ccgx-dmc/fu-ccgx-dmc-plugin.h
@@ -1,0 +1,11 @@
+/*
+ * Copyright (C) 2022 Richard Hughes <richard@hughsie.com>
+ *
+ * SPDX-License-Identifier: LGPL-2.1+
+ */
+
+#pragma once
+
+#include <fwupdplugin.h>
+
+G_DECLARE_FINAL_TYPE(FuCcgxDmcPlugin, fu_ccgx_dmc_plugin, FU, CCGX_DMC_PLUGIN, FuPlugin)
