+++ plugins/ccgx-dmc/fu-ccgx-dmc.rs
