+++ plugins/ccgx/ccgx.quirk
@@ -50,95 +50,3 @@ Summary = CCGx Power Delivery Device (Symmetric FW1)
 
 [USB\VID_04B4&PID_5218&SID_1F00&APP_6432&MODE_FW2]
 Summary = CCGx Power Delivery Device (Symmetric FW2)
-
-# Lenovo ThinkPad Universal USB-C Dock
-[USB\VID_17EF&PID_30A9]
-Plugin = ccgx
-GType = FuCcgxDmcDevice
-Summary = Dock Management Controller Device
-ParentGuid = USB\VID_17EF&PID_30AF
-Name = ThinkPad Universal USB-C Dock
-Flags = has-manual-replug
-CcgxDmcTriggerCode = 0x02
-InstallDuration = 60
-
-[USB\VID_17EF&PID_3105]
-Plugin = ccgx
-GType = FuCcgxDmcDevice
-Summary = Dock Management Controller Device
-ParentGuid = USB\VID_17EF&PID_30AF
-Name = ThinkPad Universal USB-C Dock
-Flags = has-manual-replug
-CcgxDmcTriggerCode = 0x02
-InstallDuration = 60
-
-# HP USB-C Dock G5
-[USB\VID_03F0&PID_046B]
-Plugin = ccgx
-GType = FuCcgxDmcDevice
-Summary = Dock Management Controller Device
-ParentGuid = USB\VID_03F0&PID_0363
-Vendor = HP
-Name = USB-C Dock G5
-CcgxDmcTriggerCode = 0x01
-InstallDuration = 233
-RemoveDelay = 203000
-
-# HP USB-C/A Universal Dock G2
-[USB\VID_03F0&PID_0A6B]
-Plugin = ccgx
-GType = FuCcgxDmcDevice
-Summary = Dock Management Controller Device
-ParentGuid = USB\VID_03F0&PID_096B
-Vendor = HP
-Name = USB-C/A Universal Dock G2
-CcgxDmcTriggerCode = 0x01
-InstallDuration = 180
-RemoveDelay = 162000
-
-# HP Thunderbolt Dock G4 Root Hub
-[USB\VID_1D5C&PID_5801]
-Summary = USB Hub
-Name = Thunderbolt Dock G4 top most USB Hub
-
-# HP Thunderbolt Dock G4 Hub
-[USB\VID_03F0&PID_2488]
-Summary = USB Hub
-ParentGuid = USB\VID_1D5C&PID_5801
-Name = Thunderbolt Dock G4 USB Hub
-
-# HP Thunderbolt Dock G4
-[USB\VID_03F0&PID_0488]
-Plugin = ccgx
-GType = FuCcgxDmcDevice
-Summary = Dock Management Controller Device
-ParentGuid = USB\VID_03F0&PID_2488
-Vendor = HP
-Name = Thunderbolt Dock G4
-CcgxDmcTriggerCode = 0x01
-InstallDuration = 898
-RemoveDelay = 732000
-
-# Quanta Storage Inc. QSI Thunderbolt4 Godzilla Hub
-[USB\VID_2BEF&PID_9065]
-Plugin = ccgx
-GType = FuCcgxDmcDevice
-Summary = Dock Management Controller Device
-CcgxDmcTriggerCode = 0x01
-RemoveDelay = 732000
-
-# Anker Thunderbolt4 Mini Dock
-[USB\VID_291A&PID_8398]
-Plugin = ccgx
-GType = FuCcgxDmcDevice
-Summary = Dock Management Controller Device
-CcgxDmcTriggerCode = 0x01
-RemoveDelay = 732000
-
-# Caldigit ElementHub
-[USB\VID_2188&PID_0035]
-Plugin = ccgx
-GType = FuCcgxDmcDevice
-CcgxDmcTriggerCode = 0x01
-[USB\VID_2188&PID_0035&CID_05&VER_3.3.1.69]
-CcgxDmcCompositeVersion = 15
