+++ plugins/meson.build
@@ -26,6 +26,7 @@ plugins = [
   'bcm57xx',
   'bios',
   'ccgx',
+  'ccgx-dmc',
   'cfu',
   'ch341a',
   'ch347',
